🚛 refactoring
Moving Pulumi files into /deploy/ folder
Aiming to experiment with class constructors, instead of the current inline format

My problem is that testing the index.ts file imports and creates the
entire infrastructure.

We can test smaller pieces by creating smaller components with classes
